requirements.txt
Flask==2.2.2
gunicorn
Werkzeug==2.2.2
mysql-connector-python==8.3.0